requirements.txt
discord.py
emoji
pyyaml
